+++ src/main/java/com/care/project/main/service/ReviewService.java
@@ -15,5 +15,5 @@ public interface ReviewService {
 	public int getReserveCount(String id);
 	public int reviewCheck(String id, int movieid);
 	public int writeReview (ReviewDTO dto);
-	public int delReserve(int id);
+	public int delReserve(long id);
  }
